# opt: remove bound and loop unrolling in quote (begin to use clang13)