--- a/bench.py
+++ b/bench.py
@@ -70,7 +70,7 @@ def compare(args):
     run("git checkout -- .")
     if current_branch != "main":
         run("git checkout main")
-    run("git pull origin main")
+    run("git pull --allow-unrelated-histories origin main")
 
     # benchmark main branch
     (fd, main) = tempfile.mkstemp(".main.txt")