… .github/workflows/benchmark-linux-amd64.yml
… README.md
… ast/encode_test.go
… internal/native/avx2/native_amd64.s
… internal/native/avx2/native_subr_amd64.go